move directory chat
