+++ domain/src/main/java/com/jin/domain/usecase/EnsureInitialMessageUseCase.kt
@@ -1,6 +1,6 @@
 package com.jin.domain.usecase
 
-import com.jin.domain.repositories.ChatRepository
+import com.jin.domain.chat.ChatRepository
 
 class EnsureInitialMessageUseCase(private val chatRepository: ChatRepository) {
     suspend operator fun invoke(menuName: String) {
