+++ domain/src/main/java/com/jin/domain/usecase/GetMessageListUseCase.kt
@@ -1,8 +1,8 @@
 package com.jin.domain.usecase
 
 import androidx.paging.PagingData
-import com.jin.domain.repositories.ChatRepository
-import com.jin.domain.model.chat.ChatItem
+import com.jin.domain.chat.ChatRepository
+import com.jin.domain.chat.model.ChatItem
 import kotlinx.coroutines.flow.Flow
 
 class GetMessageListUseCase(private val chatRepository: ChatRepository) {
