+++ domain/src/main/java/com/jin/domain/usecase/SendMessageUseCase.kt
@@ -1,6 +1,6 @@
 package com.jin.domain.usecase
 
-import com.jin.domain.repositories.ChatRepository
+import com.jin.domain.chat.ChatRepository
 
 class SendMessageUseCase(private val chatRepository: ChatRepository) {
     suspend operator fun invoke(menuName: String, message: String) {
